+++ applications/client/src/views/Campaign/Explore/Panels/Meta/ServerMeta.tsx
@@ -11,7 +11,7 @@ import { ToggleHiddenDialog } from '@redeye/client/views';
 import { useMutation } from '@tanstack/react-query';
 import { observer } from 'mobx-react-lite';
 import { MenuItem2 } from '@blueprintjs/popover2';
-import { MetaGridLayout, MetaHeader, SaveInputButton, ToggleHiddenButton } from './Meta.styles';
+import { MetaGridLayout, MetaHeader, SaveInputButton, ToggleHiddenButton } from './MetaComponents';
 import { useToggleHidden } from '../hooks/use-toggle-hidden';
 import { useCheckLastUnhidden } from '../hooks/use-check-last-unhidden';
